Fix tests

+++ tests/test_orchestrator.py
@@ -13,6 +13,7 @@ from eva_sub_cli.exceptions.submission_status_exception import SubmissionStatusE
 from eva_sub_cli.orchestrator import orchestrate_process, VALIDATE, SUBMIT, DOCKER, check_validation_required
 from eva_sub_cli.submit import SUB_CLI_CONFIG_KEY_SUBMISSION_ID
 from eva_sub_cli.validators.validator import READY_FOR_SUBMISSION_TO_EVA
+from tests.test_utils import touch
 
 
 class TestOrchestrator(unittest.TestCase):
@@ -32,9 +33,17 @@ class TestOrchestrator(unittest.TestCase):
             shutil.rmtree(self.test_sub_dir)
         os.makedirs(self.test_sub_dir)
         shutil.copy(os.path.join(self.resource_dir, 'EVA_Submission_test.json'), self.metadata_json)
+        shutil.copy(os.path.join(self.resource_dir, 'EVA_Submission_test.xlsx'), self.metadata_xlsx)
+        for file_name in ['example1.vcf.gz', 'example2.vcf', 'example3.vcf', 'GCA_000001405.27_fasta.fa']:
+            touch(os.path.join(self.test_sub_dir, file_name))
+        self.curr_wd = os.curdir
+        os.chdir(self.test_sub_dir)
+
 
     def tearDown(self) -> None:
-        shutil.rmtree(self.test_sub_dir)
+        os.chdir(self.curr_wd)
+        if os.path.exists(self.test_sub_dir):
+            shutil.rmtree(self.test_sub_dir)
 
     def test_check_validation_required(self):
         tasks = ['submit']
@@ -213,8 +222,6 @@ class TestOrchestrator(unittest.TestCase):
 
 
     def test_orchestrate_with_metadata_xlsx(self):
-        shutil.copy(os.path.join(self.resource_dir, 'EVA_Submission_test.xlsx'), self.metadata_xlsx)
-
         with patch('eva_sub_cli.orchestrator.WritableConfig') as m_config, \
                 patch('eva_sub_cli.orchestrator.DockerValidator') as m_docker_validator:
             orchestrate_process(self.test_sub_dir, None, None, None, self.metadata_xlsx,
